requirements.txt
beautifulsoup4==4.8.0
dj-database-url==0.5.0
Django==2.2.5
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.0.1
django-datatable==0.3.1
django-datatables-view==1.19.0
django-filter==2.2.0
-e git+https://github.com/rodrigoamaral/django-fullcalendar.git@d116fcfc2093e57035932a9ea7c111d449fd7e4d#egg=django_fullcalendar
django-heroku==0.3.1
django-sass==0.2.0
djangorestframework==3.10.3
gunicorn==19.9.0
libsass==0.19.2
Markdown==3.1.1
Pillow==6.1.0
psycopg2==2.8.3
pytz==2019.2
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
whitenoise==4.1.3
